menambah menu login

+++ contoh/app2.js
@@ -4,12 +4,14 @@ const { body, validationResult } = require('express-validator')
 const session = require('express-session')
 const cookieParser = require('cookie-parser')
 const flash = require('connect-flash')
+const bcrypt = require('bcrypt')
 
 const methodOverride = require('method-override')
 
-require('../database')
-const Book = require('../model/books');
-const User = require('../model/users');
+require('./database')
+const Book = require('./model/books');
+const User = require('./model/users');
+const userRoutes = require('./routes/user.route');
 
 const app = express();
 const port = 3000;
@@ -25,127 +27,139 @@ app.use(express.static('public'))
 app.use(express.urlencoded({ extended: true }))
 
 // konfigurasi flash
+const oneDay = 1000 * 60 * 60 * 24;
 app.use(cookieParser('secret'));
 app.use(
     session({
-        cookie: { maxAge: 6000 },
+        cookie: { maxAge: oneDay},
+        // key: session,
         secret: 'secret',
-        resave: true,
+        resave: false,
         saveUninitialized: true,
     })
-);
+)
 
 app.use(flash())
 
-app.get('/', (req, res) => {
-    // res.send('hello world')
-    // res.sendFile('./index.html', {root: __dirname})
-    res.status(200)
-    res.render('index', { title: 'Halaman Home', layout: 'main-layout', })
+app.use(userRoutes)
+
+app.get('/', async (req, res) => {
+    const userLogin = await User.find({email: req.session.email})
+    if(req.session.email === undefined){
+        res.redirect('login')
+    }else{
+        res.status(200)
+        res.render('index', { title: 'Halaman Home', layout: 'layouts/main-layout', userLogin})
+    }
 })
 
-app.get('/login', (req, res) => {
-    // res.send('hello world')
-    // res.sendFile('./index.html', {root: __dirname})
-    res.status(200)
-    res.render('login', { title: 'Halaman Login', layout: 'layouts/main-layout', })
+app.get('/loginUser', (req, res) => {
+    if(req.session.email === undefined){
+        res.redirect('/books')
+    }else{
+        res.redirect('/')
+    }
 })
 
 app.post('/loginUser', [
-        body('email').custom(async(value) => {
-            const cekUser = await User.findOne({ email: value })
-            if (!cekUser) { // jika ada user 
+        body('email').custom(async (valueEmail) => { 
+            const cekUser = await User.findOne({ email: valueEmail })
+            if(valueEmail === ''){
+                throw new Error('email harus diisi')
+            }
+            else if (!cekUser) { // jika ada user 
                 throw new Error('email salah')
             }
             return true;
         }),
-        body('password').custom(async(value) => {
-            const cekPassword = await User.findOne({ password: value })
-            if (!cekPassword) { // jika password benar
+        body('password').custom( async(valuePassword, {req}) => {
+            // console.log(cekPassword);
+            const user = await User.findOne({email: req.body.email})
+            const matchPass = await bcrypt.compare(valuePassword, user.password)
+            // cocokan pasword dari ui dengan yg telah di hash di db
+            // const cekEmailNPassword = await User.findOne({$and: [{email: {$eq: req.body.email}}, {password: {$eq: valuePassword}}]})
+            // console.log(matchPass)
+            if (valuePassword === ''){
+                throw new Error('Password harus diisi')
+            }
+            if (!cekEmailNPassword) { // jika email dan password tidak cocok
                 throw new Error('Password Salah!')
             }
             return true;
         }),
     ],
-    (req, res) => {
+    async (req, res) => {
         const errors = validationResult(req)
-        console.log(errors)
+        // const userLogin = await User.findOne({email: req.body.email})
+        const books = await Book.find()
         if (!errors.isEmpty()) { // jika error request tidak kosong
             // return res.status(400).json({errors: errors.array()})
             res.render('login', {
                 title: 'Halaman Login',
-                layout: 'layouts/main-layout',
+                layout: 'login',
                 errors: errors.array()
             });
         } else {
-            req.flash('msg', 'Login Berhasil')
+            // req.flash('msg', 'Login Berhasil')
+            const session = req.session
+            session.email = req.body.email
+            const userLogin = await User.find({email: req.session.email})
+            // console.log(userLogin)
             res.status(200)
-            res.redirect('/books')
+            // res.redirect('/books')
+            if(req.session.email === undefined){
+                res.redirect('/login')
+            }else{
+                res.render('books', {books, userLogin, title: 'Halaman Buku', layout: 'layouts/main-layout', msg: 'Login Berhasil'})
+            }
                 // });
         }
     })
 
-app.get('/register', (req, res) => {
-    // res.send('hello world')
-    // res.sendFile('./index.html', {root: __dirname})
-    res.status(200)
-    res.render('register', { title: 'Halaman Register', layout: 'layouts/main-layout', })
-})
-
-app.post('/tambahUser', [
-        body('email').custom(async(value) => {
-            const duplikatEmail = await User.findOne({ email: value })
-            if (duplikatEmail) { // jika ada data nama buku yang sama
-                throw new Error('email sudah digunakan')
-            }
-            return true;
-        }),
-    ],
-    (req, res) => {
-        const errors = validationResult(req)
-        if (!errors.isEmpty()) { // jika ada error request
-            // return res.status(400).json({errors: errors.array()})
-            res.render('tambah', {
-                title: 'Tambah Data User',
-                layout: 'layouts/main-layout',
-                errors: errors.array()
-            });
-        } else {
-            // console.log(req.body)
-            User.insertMany({
-                    nama: req.body.nama,
-                    email: req.body.email,
-                    password: req.body.password,
-                    role: 2
-                },
-                (error, result) => {
-                    // kirimkan flash message
-                    req.flash('msg', 'Register Berhasil')
-                    res.status(200)
-                    res.redirect('/books')
-                });
-        }
-    })
-
-
 app.get('/books', async(req, res) => {
+    // const session = req.session
     const books = await Book.find()
-    res.status(200)
-    res.render('books', { title: 'Halaman Buku', layout: 'layouts/main-layout', books, msg: req.flash('msg') })
+    const userLogin = await User.find({email: req.session.email})
+    // console.log(userLogin)
+    if(req.session.email === undefined){
+        res.redirect('/login')
+    }else{
+        res.status(200)
+        res.render('books', { title: 'Halaman Buku', layout: 'layouts/main-layout', books, userLogin, msg: req.flash('msg') })
+    }
 })
 
 
 app.get('/users', async(req, res) => {
+    const userLogin = await User.find({email: req.session.email})
     const users = await User.find()
-    res.status(200)
-    res.render('users', { title: 'Halaman User', layout: 'layouts/main-layout', users, msg: req.flash('msg') })
+    // console.log(userLogin)
+    if(userLogin[0].role !== 1){
+        res.redirect('/')
+    } else{
+        if(req.session.email === undefined){
+            res.redirect('/login')
+        }else{
+            res.status(200)
+            res.render('users', { title: 'Halaman User', layout: 'layouts/main-layout', userLogin, users, msg: req.flash('msg') })
+        }
+    }
 })
 
 // mengubah data buku
 app.get('/ubah/:namaBuku', async(req, res) => {
     const book = await Book.findOne({ namaBuku: req.params.namaBuku })
-    res.status(200)
-    res.render('ubah', { title: 'Ubah Data Buku', layout: 'layouts/main-layout', book })
+    const userLogin = await User.find({email: req.session.email})
+    if(req.session.email === undefined){
+        res.redirect('/login')
+    }else{
+        if(userLogin[0].role === 3){
+            res.redirect('/books')
+        }else{
+            res.status(200)
+            res.render('ubah', { title: 'Ubah Data Buku', layout: 'layouts/main-layout', book, userLogin })
+        }
+    }
 })
 
 app.put('/ubah', [
@@ -161,7 +175,9 @@ app.put('/ubah', [
     ],
     (req, res) => {
         // console.log(req.body)
+        const userLogin = User.find({email: req.session.email})
         const errors = validationResult(req)
+
         if (!errors.isEmpty()) { // jika ada eror
             // return res.status(400).json({errors: errors.array()})
             res.render('ubah', {
@@ -179,15 +195,25 @@ app.put('/ubah', [
                     pengarang: req.body.pengarang
                 }
             }).then((result) => {
+                // res.render('books',{userLogin})
                 req.flash('msg', 'Data Buku Berhasil Diubah')
                 res.redirect('/books')
             });
         }
     })
 
-app.get('/tambah', (req, res) => {
-    res.status(200)
-    res.render('tambah', { title: 'Tambah Data Buku', layout: 'layouts/main-layout' })
+app.get('/tambah', async (req, res) => {
+    const userLogin = await User.find({email: req.session.email})
+    if (req.session.email === undefined){
+        res.redirect('/login')
+    } else {
+        if(userLogin[0].role === 3){
+            res.redirect('/books')
+        } else{
+            res.status(200)
+            res.render('tambah', { title: 'Tambah Data Buku', layout: 'layouts/main-layout', userLogin })
+        }
+    }
 })
 
 // proses data buku melalui form 
@@ -201,13 +227,15 @@ app.post('/tambah', [
         }),
     ],
     (req, res) => {
+        const userLogin = User.find({email: req.session.email})
         const errors = validationResult(req)
         if (!errors.isEmpty()) { // jika ada error request
             // return res.status(400).json({errors: errors.array()})
             res.render('tambah', {
                 title: 'Tambah Data Buku',
                 layout: 'layouts/main-layout',
-                errors: errors.array()
+                errors: errors.array(),
+                userLogin
             });
         } else {
             // console.log(req.body)
@@ -215,50 +243,54 @@ app.post('/tambah', [
                 // kirimkan flash message
                 req.flash('msg', 'Data Buku Berhasil Ditambahkan')
                 res.status(200)
-                res.redirect('/books')
+                res.redirect('/books', {userLogin})
             });
         }
     })
 
-app.get('/about', (req, res) => {
-    // res.send('ini adalah halaman about')
-    // res.sendFile('./about.html', {root: __dirname})
-    res.render('about', { title: 'Halaman About', layout: 'layouts/main-layout' })
+app.get('/about', async (req, res) => {
+    const userLogin = await User.find({email: req.session.email})
+    if(req.session.email === undefined){
+        res.redirect('/login')
+    }else{
+        res.render('about', { title: 'Halaman About', layout: 'layouts/main-layout', userLogin })
+    }
 })
 
-// app.get('/product/:id', (req, res) => {
-//     res.send(`Product id : ${req.params.id} <br> category: ${req.query.category}`)
-// })
-
-
-// menghapus data buku
-// app.get('/hapus/:namaBuku', async (req, res) => {
-//     const book = await Book.findOne({namaBuku: req.params.namaBuku})
-//     if(!book){ // jika namaBuku tidak ada 
-//         res.status(404)
-//         res.render('page_error', {title: 'Nama Buku tidak ada', layout: 'layouts/main-layout'})
-//     }else{ // jika ada namaBuku
-//         Book.deleteOne({_id: book._id}).then( (result) => {
-//             req.flash('msg', 'Data Buku Berhasil Dihapus')
-//             res.redirect('/books')
-//         })
-//     }
-// })
-
 // menghapus data buku opsi kedua
-app.delete('/hapus', (req, res) => {
-    // res.send(req.body)
-    Book.deleteOne({ namaBuku: req.body.namaBuku }).then((result) => {
+app.delete('/hapus', async (req, res) => { //hapus?_method=DELETE
+    const userLogin = await User.find({email: req.session.email})
+    if(userLogin[0].role === 3){
+        res.redirect('/books')
+    } else{
+    
+        Book.deleteOne({ namaBuku: req.body.namaBuku }).then((result) => {
         req.flash('msg', 'Data Buku Berhasil Dihapus')
         res.status(200)
         res.redirect('/books')
     });
+        // res.status(200)
+        // res.render('books', {title: 'Halaman Buku', layout: 'layouts/main-layout', msg: 'Data Buku Berhasil Dihapus', userLogin, books})
+    }
 });
 
-app.use('/', (req, res) => { // untuk menangkap url yang tidak ada
-    res.status(404)
-    res.render('page_error', { title: 'Halaman Tidak Ditemukan', layout: 'layouts/main-layout' })
+app.get('/logout',(req,res) => {
+    req.session.destroy();
+    res.redirect('/login');
+});
+
+app.use('/', async (req, res) => { // untuk menangkap url yang tidak ada
+    const userLogin = await User.find({email: req.session.email})
+    // console.log(userLogin)
+    if(req.session.email === undefined){
+        res.redirect('/login')
+    }else{
+        res.status(404)
+        res.render('page_error', { title: 'Halaman Tidak Ditemukan', layout: 'layouts/main-layout', userLogin })
+        // res.redirect('/page_error')
+    }
 })
+
 app.listen(port, () => {
     console.log(`Books App | listening at http://localhost:${port}`)
 });
